# perf(imports): speed up imports by at least 2x