--- a/.github/workflows/ibis-backends.yml
+++ b/.github/workflows/ibis-backends.yml
@@ -665,6 +665,7 @@ jobs:
               - mssql
             extras:
               - mssql
+              - geospatial
           - name: mysql
             title: MySQL
             services: