--- a/ibis/backends/bigquery/client.py
+++ b/ibis/backends/bigquery/client.py
@@ -4,9 +4,6 @@ from __future__ import annotations
 
 import functools
 
-import google.cloud.bigquery as bq
-import pandas as pd
-
 import ibis.common.exceptions as com
 import ibis.expr.datatypes as dt
 import ibis.expr.operations as ops
@@ -73,6 +70,8 @@ def bigquery_param(dtype, value, name):
 
 @bigquery_param.register
 def bq_param_struct(dtype: dt.Struct, value, name):
+    import google.cloud.bigquery as bq
+
     fields = dtype.fields
     field_params = [bigquery_param(fields[k], v, k) for k, v in value.items()]
     result = bq.StructQueryParameter(name, *field_params)
@@ -98,12 +97,18 @@ def bq_param_array(dtype: dt.Array, value, name):
             raise TypeError("ARRAY<ARRAY<T>> is not supported in BigQuery")
         else:
             query_value = value
+
+        import google.cloud.bigquery as bq
+
         result = bq.ArrayQueryParameter(name, bigquery_type, query_value)
         return result
 
 
 @bigquery_param.register
 def bq_param_timestamp(_: dt.Timestamp, value, name):
+    import google.cloud.bigquery as bq
+    import pandas as pd
+
     # TODO(phillipc): Not sure if this is the correct way to do this.
     timestamp_value = pd.Timestamp(value, tz="UTC").to_pydatetime()
     return bq.ScalarQueryParameter(name, "TIMESTAMP", timestamp_value)
@@ -111,26 +116,37 @@ def bq_param_timestamp(_: dt.Timestamp, value, name):
 
 @bigquery_param.register
 def bq_param_string(_: dt.String, value, name):
+    import google.cloud.bigquery as bq
+
     return bq.ScalarQueryParameter(name, "STRING", value)
 
 
 @bigquery_param.register
 def bq_param_integer(_: dt.Integer, value, name):
+    import google.cloud.bigquery as bq
+
     return bq.ScalarQueryParameter(name, "INT64", value)
 
 
 @bigquery_param.register
 def bq_param_double(_: dt.Floating, value, name):
+    import google.cloud.bigquery as bq
+
     return bq.ScalarQueryParameter(name, "FLOAT64", value)
 
 
 @bigquery_param.register
 def bq_param_boolean(_: dt.Boolean, value, name):
+    import google.cloud.bigquery as bq
+
     return bq.ScalarQueryParameter(name, "BOOL", value)
 
 
 @bigquery_param.register
 def bq_param_date(_: dt.Date, value, name):
+    import google.cloud.bigquery as bq
+    import pandas as pd
+
     return bq.ScalarQueryParameter(
         name, "DATE", pd.Timestamp(value).to_pydatetime().date()
     )