--- a/ibis/backends/dask/__init__.py
+++ b/ibis/backends/dask/__init__.py
@@ -17,7 +17,6 @@ import ibis.expr.types as ir
 from ibis.backends.dask.core import execute_and_reset
 from ibis.backends.pandas import BasePandasBackend
 from ibis.backends.pandas.core import _apply_schema
-from ibis.formats.pandas import DaskData
 
 # Make sure that the pandas backend options have been loaded
 ibis.pandas  # noqa: B018
@@ -111,6 +110,8 @@ class Backend(BasePandasBackend):
         return execute_and_reset(query.op(), params=params, **kwargs)
 
     def table(self, name: str, schema: sch.Schema = None):
+        from ibis.formats.pandas import DaskData
+
         df = self.dictionary[name]
         schema = schema or self.schemas.get(name, None)
         schema = DaskData.infer_table(df, schema=schema)