… ibis/backends/duckdb/registry.py
… ibis/backends/impala/__init__.py
… ibis/backends/impala/client.py
… ibis/backends/pandas/__init__.py
… ibis/backends/polars/__init__.py
… ibis/backends/pyspark/__init__.py
… ibis/backends/pyspark/ddl.py
… ibis/backends/snowflake/__init__.py
… ibis/backends/snowflake/datatypes.py
… ibis/backends/snowflake/registry.py
… ibis/backends/tests/test_client.py
… ibis/backends/trino/__init__.py
… ibis/backends/trino/datatypes.py
… ibis/common/caching.py
… ibis/common/temporal.py
… ibis/common/validators.py
… ibis/examples/__init__.py
… ibis/examples/tests/test_examples.py
… ibis/expr/datatypes/core.py
… ibis/expr/datatypes/value.py
… ibis/expr/operations/generic.py
… ibis/expr/sql/__init__.py
… ibis/expr/sql/core.py
… ibis/expr/types/core.py
… ibis/expr/types/relations.py
… ibis/tests/benchmarks/test_benchmarks.py
… ibis/util.py
… poetry.lock
… pyproject.toml
… requirements.txt