This PR speeds up importing the individual backends by at least 2x and in the
clickhouse case 3x+.

This shows up for users as additional unnecessary overhead when connecting to
databases due entry point loading on module attribute access.

I noticed that the time to call `ibis.duckdb.connect()`, was much much slower
than calling `duckdb.connect()`.

I expected ibis to be a bit slower, but not **27x** slower:

```
In [1]: %time import ibis; con = ibis.duckdb.connect()
CPU times: user 1.94 s, sys: 6.35 s, total: 8.29 s
Wall time: 801 ms

In [1]: %time import ibis; con = ibis.duckdb.connect()
CPU times: user 296 ms, sys: 40.1 ms, total: 336 ms
Wall time: 333 ms
```

It turned out that good chunk of time is spent importing `geoalchemy2` and
`geopandas`, just to check whether they are installed:

```
In [1]: %%time
   ...: import geopandas
   ...: import geoalchemy2
   ...:
   ...:
CPU times: user 611 ms, sys: 367 ms, total: 978 ms
Wall time: 351 ms
```

This overhead is incurred for backends that don't have geospatial support (by definition)

It turns out we are slower than necessary to import across most of our backends, mostly because
of pandas, pyarrow and numpy imports followed by some
less-expensive-to-import-but-still-non-negligible modules like sqlalchemy.

Any of the sqlalchemy backends that reuse the postgres registry were slow
because of the geospatial dependency imports.

What I did here is generally to move imports closer to where they are used/needed.

This involved a few things

1. Remove the geospatial dependency checking from the import process. We'll now
 	 always register geospatial ops and errors from missing dependencies will occur
 	 when trying to convert outputs to GeoDataFrames, or perform other Python-side
 	 geospatial operations or type conversions.
2. Move imports that are more expensive into methods and function calls.

Additionally, `import google.cloud.bigquery` alone appears to take about 700
milliseconds, so I moved that import closer to call sites as needed:

```
In [1]: %time import google.cloud.bigquery
CPU times: user 1.84 s, sys: 6.39 s, total: 8.23 s
Wall time: 705 ms
```

Finally, `clickhouse_connect` is also very slow, so I inlined those imports too:

```
In [1]: %time import clickhouse_connect
CPU times: user 1.6 s, sys: 6.42 s, total: 8.01 s
Wall time: 491 ms
```

I've also added a benchmark so that we can keep track of import times.