@cpcloud 2023-06-13
> I'm less convinced on the larger changes here - there's a readability and maintenance cost to making certain imports local (especially common ones like sqlalchemy). Unless we're testing that a certain file or backend can be used without that import being imported, I could see these imports accidentally creeping back in.

> IMO, local imports should be used only for optional dependencies, or modules that are known to be expensive to avoid importing them unless that functionality is likely to be used. So delaying the geospatial imports makes sense to me (expensive, unlikely to be used), but delaying the import of bigquery from the bigquery backend (where it will always be needed) makes less sense.

I hear what you're saying.

All of these changes here were driven by iteration using tuna and looking at what was "reasonable" to move around and what would actually decrease import times a noticeable amount. Sometimes, something was only noticeable once the sum of the import times from a bunch of apparently "cheap to import" modules were moved to the call site.

I'd encourage you to try removing some of the commits here and see what changes. It's pretty easy to look at this code and say it's hard to read etc because a lot of moving these imports are not obviously impactful. If you remove enough of the smaller moves then the import will creep back up.

> but delaying the import of bigquery from the bigquery backend (where it will always be needed) makes less sense

I don't think that's true. You're right that it's always required if you're connecting to the BigQuery service and executing queries etc, but if you're using unbound tables and the compiler you may never need some of the imports (e.g., the storage client).

I believe this holds true for all the SQL backends.