@jcrist 2023-06-13
>  I don't think that's true. You're right that it's always required if you're connecting to the BigQuery service and executing queries etc, but if you're using unbound tables and the compiler you may never need some of the imports (e.g., the storage client).

I understand not needing the storage client, but unless you're only using `ibis.bigquery.compile`, AFAICT you'll always need `google.bigquery` imported. My issue is that making this import local makes the code a bit harder to read, and without a test ensuring `ibis.bigquery.compile` can work without importing `google.cloud.bigquery`, there's a decent chance there's a regression in this dependency management. I'm not convinced this is worth it. Same goes for `import duckdb` in the `duckdb` backend.

> I'd encourage you to try removing some of the commits here and see what changes.

The other two I pointed out I don't think have a perf effect. `rich` is a very light import, looking at that profile it seems pretty negligible compared to other imports. But making that import local seems fine if you disagree, since it's pretty manageable (though see my comment above about a behavior change).

The `sqlalchemy` one doesn't have an effect (see my comment above). For any `sqlalchemy` backend, `sqlalchemy` is already imported in the base class module. Making those imports local just muddies the code without reducing import times.