@gforsyth 2023-06-13
I'm very much of two minds on this.  

Readability counts, and Ibis is very intertwined, so some of the inlining can make it that much harder for new contributors to get up to speed.

AND

I have done horrible things in the name of improving startup times, because startup / import times do matter.

Like all parts of a user-facing library, I think it comes down to a balance of what the user experiences, and also how the maintainers are able to fix, improve, and extend the code.

We have a few categories of things here, and I think they probably each deserve their own PR / discussion, because I don't think this is a one-size-fits-all situation:

1. Geospatial-y imports -- very expensive, used by a subset of users on a subset of backends.  We should almost certainly inline these.
2. Bigquery-ish imports -- very expensive, used by almost all users of one _specific_ backend.
3. SQLAlchemical imports -- moderately expensive, used by the majority of Ibis users (supposition)
4. Pandonic imports -- very expensive, used by all users (for the moment)
5. PyArrow imports -- moderately expensive, used by most users (at least indirectly)


These each occupy a separate semantic import space.  

We can say that most users are not using the geospatial stuff, so we can (and should!) accept the loss of readability for the sake of improving performance for all other users.  

The obverse of this is bigquery imports -- we would inline those so that users who are using bigquery with unbound tables don't pay the import price.  We don't have metrics on this, but I would guess that the more common pattern is to execute those queries via Ibis (and not just using unbound tables and the compiler). In that case, we probably don't want to sacrifice readability for that subset of users.

Then we have sqlalchemy -- I think we should definitely strive to not import sqlalchemy for the backends that don't use it, but a BUNCH of backends use it, and use it everywhere, so inlining it over and over again is painful, and I think will eventually lead to a regression.

Pandas is expensive enough that it might be worth the readability-hit.

I'm unsure of how to handle pyarrow.

One thing we might try for some of the expensive but oft-used or just "very expensive" imports is https://github.com/xonsh/lazyasd -- it also comes with a readability cost, but will let us keep expensive imports at the top of a given file and avoid all the inlined imports.

I do think that we should split up this PR into components, though, to make it easier to discuss each set of changes, and also easier to measure each group of changes independently.  I know that makes it more work @cpcloud and I'm happy to take some of that so we aren't just burning all of your time.